requirements.txt
tqdm
arxiv
mistralai
python-dotenv
openpyxl
pymupdf
frontend
Pillow
torch
addict
matplotlib
torchvision
einops
transformers==4.46.3
tokenizers==0.20.3
img2pdf
easydict
numpy